Running test in python

+++ cmodules/IcsParser/src/ics_event.h
@@ -13,7 +13,7 @@ event_t getFirstEvent(ics_t *ics);
 event_t getNextEvent(ics_t *ics);
 event_t getLastEvent(ics_t *ics);
 event_t getCurrentEvent(ics_t *ics);
-size_t setCurrentEvent(ics_t *ics, size_t index);
+int setCurrentEvent(ics_t *ics, size_t index);
 event_t getEventAt(ics_t *ics, size_t index); 
 event_t getNextEventInRange(ics_t *ics, const char *start_date_str, const char *end_date_str);
 void sortEventsByStart(ics_t *ics);
